requirements.txt
blessings==1.6
bpython==0.15
curtsies==0.2.6
Django==1.9.7
django-appconf==1.0.2
django-filter==0.13.0
djangorestframework==3.3.3
greenlet==0.4.9
Markdown==2.6.6
Pillow==3.2.0
psycopg2==2.6.1
Pygments==2.1.3
pytz==2016.4
requests==2.10.0
six==1.10.0
wcwidth==0.1.6
